refactor : InOutView.kt 하나로 모여놓은 입출력 뷰 클래스를 InView,OutView 로 나누고, 그에 맞춰 테스트에 해당하는 뷰도 수정
