+++ src/main/kotlin/constant/InOutConstants.kt
@@ -1,0 +1,9 @@
+package constant
+
+object InOutConstants {
+    const val NAME = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."
+    const val COUNT = "시도할 횟수는 몇 회인가요?"
+    const val CURRENT_RESULT_FORMAT = "%s : %s"
+    const val GAME_RESULT = "\n실행 결과"
+    const val FINAL_RESULT = "최종 우승자 : "
+}
