+++ src/main/kotlin/controller/InController.kt
@@ -1,19 +1,19 @@
 package controller
 
 import constant.ErrorConstant
+import constant.InOutConstants
 import model.Car
 import model.Random
-import view.InOutConstants
-import view.InOutView
 
-class InController(private val inOutView: InOutView) {
+import view.InView
+
+class InController(private val inView: InView) {
     fun getCarName(): MutableList<Car> {
-        val input = inOutView.getData(InOutConstants.NAME)
+        val input = inView.getData(InOutConstants.NAME)
         val carNames = splitToComma(input)
         val cars = carNames.map { Car(it, randomMaker = Random()) }.toMutableList()
         return cars
     }
-
     fun splitToComma(input: String): MutableList<String> {
         val carNames = input.split(",").map { it.trim() }.filter { it.isNotBlank() }
         if (carNames.isEmpty()) throw IllegalArgumentException(ErrorConstant.ERROR_NO_NAME)
@@ -23,14 +23,13 @@ class InController(private val inOutView: InOutView) {
     }
 
     fun getTryCount(): Int {
-        val input = inOutView.getData(InOutConstants.COUNT)
+        val input = inView.getData(InOutConstants.COUNT)
         val count = parseInt(input)
         return count
     }
 
     fun parseInt(input: String): Int {
-        val count = input.toIntOrNull()
-        if (count == null) throw IllegalArgumentException(ErrorConstant.ERROR_NOT_NUMBER)
+        val count = input.toIntOrNull() ?: throw IllegalArgumentException(ErrorConstant.ERROR_NOT_NUMBER)
         if (count <= 0) throw IllegalArgumentException(ErrorConstant.ERROR_UNDER_ZERO)
         return count
     }
