+++ src/main/kotlin/view/InView.kt
@@ -1,0 +1,8 @@
+package view
+
+class InView {
+    fun getData(message: String): String {
+        println(message)
+        return readln()
+    }
+}
