+++ src/test/kotlin/InputTest.kt
@@ -2,10 +2,10 @@ import controller.InController
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
-import view.InOutView
+import view.InView
 
 class InputTest {
-    val inController = InController(InOutView())
+    private val inController = InController(InView())
 
     @Test
     fun `자동차 이름은 5자를 초과할 수 없다`() {
